[core] Get rid of the bot.session
This is due to fun.py/stats.py being the only ones that need it and referencing the global session defined from the bot is kinda painful, so I decided to make it a little easier to make HTTP requests here and use a seperate ClientSession for each cog that needs it. This might take more memory due to more sessions being defined - but it will be easier to make the requests in the future.